requirements.txt
fastapi
uvicorn
langchain
openai
faiss-cpu
python-dotenv
PyPDF2
langchain-community
pypdf
tiktoken